Expertise metier
Former-commit-id: a113df7c49b2c0a82732386b6ccb0e1a63599d55

+++ htdocs/bimpsupport/objects/BS_SAV.class.php
@@ -2258,8 +2258,8 @@ WHERE a.obj_type = 'bimp_object' AND a.obj_module = 'bimptask' AND a.obj_name = 
 			$propal = $this->getChildObject('propal');
 			$client = $this->getChildObject('client');
 
-			$encoursActu = (int) $client->getAllEncoursForSiret(true)['total'];
-			$authorisation = (int) ($client->getData('outstanding_limit') + $this->getUserLimitEncours()) * 1.2;
+			$encoursActu = $client->getAllEncoursForSiret(true)['total'];
+			$authorisation = ($client->getData('outstanding_limit') + $this->getUserLimitEncours()) * 1.2;
 			$besoin = $encoursActu + $propal->dol_object->total_ht;
 
 			if ($besoin > ($authorisation + 1)) {
